Revert "Wrapper: remove recursive bpydata symlink introduced in 2.45-2."
This reverts commit 3891882476149fddf695e7552bea7f74cb35fcdd.

Conflicts:

	debian/changelog

The setup leading to the initial commit was only due to leftovers during
upgrade path tests, and such problem hasn't been reproducible with later
upgrade path tests, so it looks like there's actually no bug.

Not rebasing interactively to erase the commit so that it remains
accessible if needed later, and so that the discussion gets linked
to: #463162.